+++ public/js/admin/admin-template-port-forward.js
@@ -1,0 +1,9 @@
+(function() {
+    if (window.__adminTemplateParts) {
+        window.__adminTemplateParts.push('\
+<div v-if="activeSection === \'port-forward\'">\
+    <port-forward-list></port-forward-list>\
+</div>\
+        ');
+    }
+})();